apps/api/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
requests
requests-oauthlib
sqlmodel
pydantic-settings
psycopg2-binary
python-dotenv
PyJWT